Resolved bugs in Admin route

+++ helperFunctions.js
@@ -1,12 +1,12 @@
 const fs = require("fs");
 
-const deletePost = (fileName, ind) => {
+const deletePost = (fileName, ind, withImage) => {
 	var data = require(`./company_data/${fileName}.json`);
 	if (ind >= data.length) {
 		res.send("Error");
 		return;
 	}
-	fs.unlinkSync(`./uploads/${data[ind].img}`);
+	if (withImage) fs.unlinkSync(`./uploads/${data[ind].img}`);
 	data.splice(ind, 1);
 	fs.writeFileSync(`./company_data/${fileName}.json`, JSON.stringify(data));
 };
@@ -29,12 +29,13 @@ const creatPost = (filename, obj, imageName) => {
 const deleteImageIfRequired = (filename, ind, company, title, image) => {
 	var data = require(`./company_data/${filename}.json`);
 	var obj = data[ind];
-	let newImageName;
-	if (obj.company != company || obj.title != title) {
-		if (image) {
+	let newImageName = " ";
+	if (obj.company !== company || obj.title !== title) {
+		if (typeof image !== "undefined") {
 			fs.unlinkSync("./uploads/" + obj.img);
 			newImageName = image.filename;
 		} else {
+			console.log("Renaming Image");
 			const str = data[ind].img;
 			const fileExtension = str.substring(str.lastIndexOf("."));
 			newImageName = `${company}_${title}${fileExtension}`;
